+++ open-sse/translator/request/openai-to-kiro.ts
@@ -5,6 +5,11 @@
 import { register } from "../registry.ts";
 import { FORMATS } from "../formats.ts";
 import { v4 as uuidv4, v5 as uuidv5 } from "uuid";
+import {
+  parseToolInput,
+  normalizeKiroToolSchema,
+  serializeToolResultContent,
+} from "./openai-to-kiro/messageHelpers.ts";
 
 /**
  * Anthropic's direct-provider `[1m]` context-1m beta suffix. Kiro is AWS
@@ -23,116 +28,8 @@ export const KIRO_UNSUPPORTED_CONTEXT_1M_MESSAGE =
  */
 export function hasUnsupportedKiroContextSuffix(model: unknown): boolean {
   return (
-    typeof model === "string" &&
-    model.toLowerCase().includes(KIRO_UNSUPPORTED_CONTEXT_1M_SUFFIX)
+    typeof model === "string" && model.toLowerCase().includes(KIRO_UNSUPPORTED_CONTEXT_1M_SUFFIX)
   );
-}
-
-function parseToolInput(value: unknown) {
-  if (value && typeof value === "object" && !Array.isArray(value)) {
-    return value;
-  }
-  if (typeof value !== "string") {
-    return {};
-  }
-
-  const trimmed = value.trim();
-  if (!trimmed) {
-    return {};
-  }
-
-  try {
-    const parsed = JSON.parse(trimmed);
-    return parsed && typeof parsed === "object" && !Array.isArray(parsed) ? parsed : {};
-  } catch {
-    return {};
-  }
-}
-
-/**
- * Recursively sanitize JSON Schema for Kiro API.
- * Kiro returns 400 "Improperly formed request" if:
- * - `required` is an empty array []
- * - `additionalProperties` is present anywhere
- */
-function normalizeKiroToolSchema(schema: unknown): Record<string, unknown> {
-  if (!schema || typeof schema !== "object" || Array.isArray(schema)) {
-    return { type: "object", properties: {} };
-  }
-
-  const result: Record<string, unknown> = {};
-  const src = schema as Record<string, unknown>;
-
-  for (const [key, value] of Object.entries(src)) {
-    // Skip empty required arrays — Kiro rejects them
-    if (key === "required" && Array.isArray(value) && value.length === 0) {
-      continue;
-    }
-    // Skip additionalProperties — Kiro doesn't support it
-    if (key === "additionalProperties") {
-      continue;
-    }
-    // Recursively process nested objects
-    if (
-      key === "properties" &&
-      typeof value === "object" &&
-      value !== null &&
-      !Array.isArray(value)
-    ) {
-      const sanitizedProps: Record<string, unknown> = {};
-      for (const [propName, propValue] of Object.entries(value as Record<string, unknown>)) {
-        sanitizedProps[propName] = normalizeKiroToolSchema(propValue);
-      }
-      result[key] = sanitizedProps;
-    } else if (typeof value === "object" && value !== null && !Array.isArray(value)) {
-      result[key] = normalizeKiroToolSchema(value);
-    } else if (Array.isArray(value)) {
-      result[key] = value.map((item) =>
-        typeof item === "object" && item !== null && !Array.isArray(item)
-          ? normalizeKiroToolSchema(item)
-          : item
-      );
-    } else {
-      result[key] = value;
-    }
-  }
-
-  return result;
-}
-
-function serializeToolResultContent(content: unknown): string {
-  if (typeof content === "string") {
-    return content || "(no output)";
-  }
-  if (!Array.isArray(content)) {
-    if (content !== null && content !== undefined) {
-      try {
-        return JSON.stringify(content);
-      } catch {
-        return "(no output)";
-      }
-    }
-    return "(no output)";
-  }
-  const parts: string[] = [];
-  for (const block of content as Array<Record<string, unknown>>) {
-    if (!block || typeof block !== "object") continue;
-    if (block.type === "text" && typeof block.text === "string") {
-      if (block.text) parts.push(block.text);
-    } else if (block.type === "image" || block.type === "image_url") {
-      const src = block.source as Record<string, unknown> | undefined;
-      const mediaType = src?.media_type ?? block.media_type ?? "image";
-      parts.push(`[image: ${mediaType}]`);
-    } else {
-      try {
-        const str = JSON.stringify(block);
-        if (str && str !== "{}") parts.push(str);
-      } catch {
-        // skip unserializable block
-      }
-    }
-  }
-  return parts.join("\n") || "(no output)";
 }
 
 /**
@@ -806,9 +703,7 @@ export function buildKiroPayload(model, body, stream, credentials) {
   // compressContext runs). This keeps conversationId stable even when compression alters content.
   // Priority 2: Deterministic hash from first user message in translated history (fallback).
   const preCompressionBody = credentials?._preCompressionBody as
-    | Record<string, unknown>
-    | null
-    | undefined;
+    Record<string, unknown> | null | undefined;
   const preCompressionMessages = Array.isArray(preCompressionBody?.messages)
     ? preCompressionBody.messages
     : null;
